Home work 21
